+++ lib/request.js
@@ -5,7 +5,7 @@ const {
   pickAll,
   pipe,
 } = require('ramda')
-const { bufferArrayToObject } = require('./helpers')
+const { bufferArrayToObject } = require('./utils')
 
 const createLanguidReq = httpReq => pipe(
   assoc('body', bufferArrayToObject(httpReq.body)),
